nats_python_template/requirements.txt
nats-py>=2.7,<3.0
python-dotenv>=1.0
nkeys>=0.2  # optional, only needed for raw jwt + nkey auth